pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Backed for CensorTracker"
authors = ["Roskomsvoboda"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "==3.0.5"
djangorestframework = "*"
django-split-settings = "*"
requests = "*"
psycopg2-binary = "*"
sentry-sdk = "*"
bcrypt = "*"
django-ipware = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
